efficientnetV2_for_classification/requirements.txt
matplotlib==3.4.2
numpy==1.21.5
opencv_python==4.5.5.64
pandas==1.4.2
Pillow==9.1.0
torch==1.9.0
torchvision==0.11.3
tqdm==4.64.0
